Separate concerns between rendering logic and ui modal
