+++ js/ui/modals/input.js
@@ -1,6 +1,7 @@
 import { warningMsg, errorMsg } from "../../lib/utils/messages.js";
-import { getFileName, clearAllEventData } from "./globals.js";
 import {
+  getFileName,
+  clearAllEventData,
   setFileData,
   getFileData,
   setFileName,
@@ -9,38 +10,17 @@ import {
   setCurrentView,
   getCurrentView,
   getCurrentEventIndex,
-  getCurrentEventNumber,
-  getCurrentEventName,
-  eventCollection,
-  currentVisObjects,
-  getCurrentVisObjects,
   saveCurrentScrollPosition,
-  getSavedScrollPosition,
 } from "../../state/globals.js";
 import {
   startPixi,
   isPixiRunning,
   getViewportPosition,
-  setViewportPosition,
-  saveSize,
 } from "../../state/pixi-state.js";
 import { hideDeploySwitch } from "../toggle/switch-deploy.js";
 import { possibleViews } from "../../viz/views/viewsDictionary.js";
 import { selectViewInformation } from "./information.js";
-import { loadObjects } from "../../loaders/load.js";
-import { objectTypes } from "../../lib/constants/objectTypes.js";
-import { copyObject } from "../../lib/utils/copy.js";
-import { checkEmptyObject } from "../../lib/utils/empty-object.js";
-import {
-  emptyViewMessage,
-  hideEmptyViewMessage,
-  showMessage,
-} from "../../lib/utils/messages.js";
-import { showViewInformation, hideViewInformation } from "./information.js";
-import { renderObjects } from "../../viz/draw/render.js";
-import { setRenderable } from "../../viz/draw/renderable.js";
-import { handleFilters } from "../filters/filter.js";
-import { setupToggles } from "../toggle/toggle.js";
+import { renderEvent } from "../../viz/renderView.js";
 
 export function hideInputModal() {
   document.getElementById("input-modal").style.display = "none";
@@ -69,111 +49,6 @@ export function clearInputModal() {
   document.getElementById("available-views").replaceChildren();
   document.getElementById("input-modal-view-selector").style.display = "none";
 }
-
-function renderEvent(eventIndex) {
-  setCurrentEventIndex(eventIndex);
-
-  // load selected event
-  const currentEventIndex = getCurrentEventIndex();
-  if (eventCollection[currentEventIndex] === undefined) {
-    const objects = loadObjects(getFileData(), getCurrentEventNumber());
-
-    eventCollection[currentEventIndex] = objects;
-
-    for (const datatype in eventCollection[currentEventIndex].datatypes) {
-      const classType = objectTypes[datatype];
-      const collection =
-        eventCollection[currentEventIndex].datatypes[datatype].collection;
-      classType.setup(collection);
-    }
-    copyObject(objects, currentVisObjects);
-  } else {
-    copyObject(eventCollection[currentEventIndex], currentVisObjects);
-  }
-
-  // update event number
-  if (eventNumberElem.firstChild) {
-    eventNumberElem.removeChild(eventNumberElem.firstChild);
-  }
-  eventNumberElem.appendChild(document.createTextNode(getCurrentEventName()));
-
-  drawView(getCurrentView());
-}
-
-async function renderView(layoutFunction, objects) {
-  const empty = checkEmptyObject(objects);
-
-  if (empty) {
-    showMessage("No objects satisfy the filter options");
-    return;
-  }
-
-  let [width, height] = layoutFunction(objects);
-  if (width === 0 && height === 0) {
-    showMessage("No objects satisfy the filter options");
-    return;
-  }
-
-  width = Math.max(width, window.innerWidth);
-  height = Math.max(height, window.innerHeight);
-  saveSize(width, height);
-  await renderObjects(objects);
-}
-
-export const drawView = async (view) => {
-  const {
-    selectorFunction,
-    layoutFunction,
-    positionFunction,
-    collections,
-    description,
-  } = possibleViews[view];
-
-  const allVisObjects = getCurrentVisObjects();
-
-  const viewObjects = {};
-  selectorFunction(allVisObjects, viewObjects);
-
-  // paint buttons
-  const buttons = document.querySelectorAll("#view-selector .view-button");
-  for (const button of buttons) {
-    if (button.innerText === view) {
-      button.style.backgroundColor = "#c5c5c5";
-    } else {
-      button.style.backgroundColor = "#f1f1f1";
-    }
-  }
-
-  const isEmpty = checkEmptyObject(viewObjects);
-
-  if (isEmpty) {
-    emptyViewMessage();
-    hideViewInformation();
-    return;
-  }
-
-  showViewInformation(view, description);
-
-  // set info button name
-  const button = document.getElementById("view-information-button");
-  button.innerText = getCurrentView();
-
-  hideEmptyViewMessage();
-
-  await renderView(layoutFunction, viewObjects);
-
-  const savedPosition = getSavedScrollPosition();
-  if (savedPosition) {
-    setViewportPosition(savedPosition.x, savedPosition.y);
-  } else {
-    positionFunction();
-    saveCurrentScrollPosition(getViewportPosition());
-  }
-
-  setRenderable(viewObjects);
-  handleFilters(viewObjects, collections, setRenderable);
-  setupToggles(collections, viewObjects);
-};
 
 // Page updates
 const eventNumberElem = document.getElementById("selected-event");
